wip on refactoring homepage
