+++ packages/web/src/app/[domain]/components/homepage/agenticSearchInfo.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+export const AgenticSearchInfo = () => {
+    return (
+        <div className="flex flex-col items-center w-fit gap-6">
+            <p>todo</p>
+        </div>
+    )
+}
